refactor: restructure project to monorepo
