+++ client/index.jsx
@@ -1,6 +1,5 @@
-import React from 'react'
 import ReactDOM from 'react-dom/client'
-import App from './App.jsx'
-import './index.css'
+import App from './components/App.jsx'
+import './assets/index.css'
 
 ReactDOM.createRoot(document.getElementById('root')).render(<App />)
