+++ webpack.config.js
@@ -1,0 +1,27 @@
+import path from 'path'
+import HtmlWebpackPlugin from 'html-webpack-plugin'
+
+export default {
+  entry: './index.js',
+  mode: 'development',
+  plugins: [
+    new HtmlWebpackPlugin()
+  ],
+  output: {
+    filename: 'bundle.js',
+    path: path.resolve(import.meta.url, 'dist')
+  },
+  module: {
+    rules: [
+      {
+        test: /\.css$/i,
+        use: ['style-loader', 'css-loader']
+      }
+    ]
+  },
+  devServer: {
+    static: './index.html',
+    compress: true,
+    port: 3001
+  }
+}
